src/pdf_generator.py
"""
PDF generator module for NZ Flyfishing Web Scraper.
Article 7 Compliance: Template-driven, database-only (no live-scraping).
"""

from pathlib import Path
from typing import Optional

from jinja2 import Environment, FileSystemLoader
from reportlab.lib.pagesizes import A4
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer
from reportlab.lib.styles import getSampleStyleSheet

from .exceptions import PDFError
from .logger import ScraperLogger
from .storage import Storage


class PDFGenerator:
    """Template-driven PDF generator for river data."""

    def __init__(self, config, logger: ScraperLogger):
        """
        Initialize PDF generator.

        Args:
            config: Configuration instance
            logger: Logger instance
        """
        self.config = config
        self.logger = logger

        # Setup Jinja2 template environment
        template_dir = config.pdf_config.get("template_dir", "templates/")
        self.jinja_env = Environment(loader=FileSystemLoader(template_dir))

    def generate_river_pdf(self, river_id: int, filename: str, storage: Storage) -> str:
        """
        Generate PDF for a single river (Article 7.2).

        Implementation stub - will be completed in Phase 8 (US6).

        Args:
            river_id: Database ID of river
            filename: Output PDF filename
            storage: Storage instance for querying data

        Returns:
            Path to generated PDF file

        Raises:
            PDFError: If generation fails or river not found
        """
        # Stub implementation
        raise PDFError("PDF generation not yet implemented")

    def generate_batch_pdfs(self, region_id: int, output_zip: str, storage: Storage) -> str:
        """
        Generate PDFs for all rivers in a region.

        Implementation stub - will be completed in Phase 8 (US6).

        Args:
            region_id: Database ID of region
            output_zip: Output ZIP filename
            storage: Storage instance

        Returns:
            Path to ZIP file
        """
        # Stub implementation
        raise PDFError("Batch PDF generation not yet implemented")

    def render_template(self, river_id: int, storage: Storage) -> str:
        """
        Render HTML template for debugging.

        Args:
            river_id: Database ID of river
            storage: Storage instance

        Returns:
            Rendered HTML as string
        """
        # Stub implementation
        return "<html><body>Template not yet implemented</body></html>"

    def close(self):
        """Cleanup resources."""
        pass
